% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTestTheory.R
\name{addAlpha}
\alias{addAlpha}
\title{Add Cronbach Alpha to MC Test Data Object}
\usage{
addAlpha(mctd)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}
}
\description{
Runs \code{\link[psych]{alpha}} from the \link{psych} package and adds
Cronbach alpha and overall student test score results to
\link{mcTestAnalysisData} as \code{alpha} and \code{scores}.
}

