% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTestTheory.R
\name{addDiscriminationIndex}
\alias{addDiscriminationIndex}
\title{Calculate Conventional Item Discrimination Index}
\usage{
addDiscriminationIndex(mctd, percentile = 0.27)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}

\item{percentile}{Percentile defining upper and lower groups (standard is
0.27)}
}
\description{
Calculates the conventional item discrimination index, based on complete test
results only. Follows the formulation described in Lopez (1998), where item
discrimination \eqn{d = p(UG) - p(LG)}. UG and LG are the upper and lower
groups -- the upper and lower 27\% of students as ranked by their overall
test scores -- and \eqn{p(UG)} and \eqn{p(LG)} are the proportion of correct
answers in each group.
}
\references{
\emph{The Item Discrimination Index: Does it Work?} Tristan Lopez
  A. Rasch Measurement Transactions, 1998, 12:1 p. 626.
  \url{http://www.rasch.org/rmt/rmt121r.htm}
}
