% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemAnalysis.R
\name{addItemAnalysis}
\alias{addItemAnalysis}
\title{Add item analysis to MC Test Data Object}
\usage{
addItemAnalysis(mctd)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}
}
\description{
Runs \code{\link{item.exam}} from the \link{psychometric} package and adds
item analysis results to \link{mcTestAnalysisData} as \item{item.analysis}.
}

