% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTestTheory.R
\name{addItemScore}
\alias{addItemScore}
\title{Add item score to MC Test Data Object}
\usage{
addItemScore(mctd)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}
}
\description{
Adds a matrix called \code{item.score} to the \link{mcTestAnalysisData}
object. Rows are students, columns represent questions and each entry is
a binary correct/incorrect response.
}
