% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemAnalysis.R
\name{addSubscaleConcept}
\alias{addSubscaleConcept}
\title{Calculate Subscale Alpha and Scores}
\usage{
addSubscaleConcept(mctd)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}
}
\description{
Add subscale alpha and CTT reliability analysis to \link{mcTestAnalysisData}.
}

