% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemResponseTheory.R
\name{addTetrachoric}
\alias{addTetrachoric}
\title{Add Tetrachoric Correlation}
\usage{
addTetrachoric(mctd)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}
}
\description{
Adds tetrachoric correlation to the \link{mcTestAnalysisData} object.
}
