% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReport.R
\name{createReport}
\alias{createReport}
\title{Create MC Test Analysis Report}
\usage{
createReport(answer_file = file.choose(), test_file = file.choose(),
  test_title = ifelse(interactive(), readline("Test title: "),
  "MC Test Analysis"), author = ifelse(interactive(),
  readline("Test author: "), Sys.info()["user"]), out_file = paste0("Report",
  " - ", test_title), out_path = getwd(), out_fmt = "pdf",
  open_file = TRUE, verbose = TRUE, report_options = list(), ...)
}
\arguments{
\item{answer_file}{Path to the answer key data file}

\item{test_file}{Path to the test results data file}

\item{test_title}{Title of the test being analyzed}

\item{author}{Name of the test administrator(s)}

\item{out_file}{Desired name of report file when saved}

\item{out_path}{Directory where report file should be saved, defaults to
current working directory}

\item{out_fmt}{Either \code{'pdf'} or \code{'html'}, indicating format of
report file. PDF files will be smaller and better typset, HTML files may be
easier to navigate. Note that \code{pdflatex} is required to be installed.
If \code{pdflatex} is not found, the report will fall back to HTML output.}

\item{open_file}{Should the report file be opened automatically once
generated?}

\item{verbose}{If TRUE, notes about progress are written to the console.}

\item{report_options}{A list of options that affect analysis parameters and
results displayed inside the report. Report options include
\code{irt_model_choice}, \code{icc_group} (\code{"question"} or
\code{"concept"}), \code{efa.nfactors}, \code{efa.cut}, \code{efa.rotate},
\code{efa.fm} and \code{distractor.pct}.}

\item{...}{Arguments passed to \code{\link{read.csv}}}
}
\description{
Creates a static PDF file containing a report on the properties of the test
using Classic Test Theory and Item Response Theory.
}
