% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemAnalysis.R
\name{discrimination_difficulty_plot}
\alias{discrimination_difficulty_plot}
\title{Discrimination To Plot Data}
\usage{
discrimination_difficulty_plot(mctd, percentile = 0.27)
}
\description{
From Alvaro: We are getting discrimination index with this function. This
function will allow you to obtain the discrimination index at the selected
percentile once the code runs.
}
\keyword{internal}

