% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemAnalysis.R
\name{discriminationDifficultyPlot}
\alias{discriminationDifficultyPlot}
\title{Discrimination To Plot Data}
\usage{
discriminationDifficultyPlot(mctd, type = "conventional",
  show_labels = TRUE, hide_legend = TRUE, show_guidelines = TRUE,
  max_limits = "max_x")
}
\description{
From Alvaro: We are getting discrimination index with this function. This
function will allow you to obtain the discrimination index at the selected
percentile once the code runs.
}

