% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distractor.R
\name{distractorPlot}
\alias{distractorPlot}
\title{Plot Distractor Analysis}
\usage{
distractorPlot(mctd, pct = 0.33, pct_relative = FALSE, use_title = FALSE)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}

\item{pct}{Percentage for top/bottom comparison}

\item{pct_relative}{Should relative (within-group) percentage or overall
percentage be displayed?}

\item{use_title}{Should the question title be included in the plot output?}
}
\description{
Plots the distractor analysis results as provided by
\code{\link{distractorTable}}.
}
