% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distractor.R
\name{distractorTable}
\alias{distractorTable}
\title{Distractor Analysis}
\usage{
distractorTable(mctd, pct = 0.33)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}

\item{pct}{Percentage for top/bottom comparison}
}
\description{
Respondents are grouped into High- or Low-performing quantiles according to
their overall test score and the quantile given by \code{pct}. If a middle
group exists, it is ignored. Note that the two groups may have differing
sizes depending on the number of respondents, questions and the distribution
of test scores. A table is returned containing a count, within-group
percentage and overall percentage of respondents from each of the performance
groups who chose each option for all questions.
}

