% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorAnalysis.R
\name{efaTable}
\alias{efaTable}
\title{Exploratory Factor Table}
\usage{
efaTable(mctd, cut = 0.3, ...)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}

\item{...}{Passed to \code{\link{addEFA}}.}
}
\description{
Prints EFA factor loadings.
}

