% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_example_data.R
\name{generateExampleData}
\alias{generateExampleData}
\title{Generate Example Test Data}
\usage{
generateExampleData(n.students = 100, n.concepts = 4, n.q_per_concept = 5,
  concept_difficulty_min_max = c(0.25, 1),
  questions_difficulty_mean_sd = c(-0.5, 0.75),
  students_overall_ability_mean_sd = c(0.25, 0.5), seed = 42,
  write_files = TRUE)
}
\arguments{
\item{n.students}{Number of "students"}

\item{n.concepts}{Number of "concepts" being tested}

\item{n.q_per_concept}{Number of questions per concept group}

\item{concept_difficulty_min_max}{Each concept is assigned a difficulty from
a uniform distribution with min and max as defined by this parameter.}

\item{questions_difficulty_mean_sd}{Each question has a difficult level that
is sampled from a normal distribution with mean and standard deviation as
defined by this parameter.}

\item{students_overall_ability_mean_sd}{Each student has an overall ability
that is used to scale "his/her" performance across the entire test.}

\item{seed}{Sets the for this run, for reproducibility}

\item{write_files}{Should .csv files be written to the working directory?}
}
\description{
This is a quick-and-dirty style exam results generator that can be used to
provide, if nothing else, an example test data set that will work throughout
all of the MCTestAnalysis functions and interfaces.
}

