% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{loadDataGadget}
\alias{loadDataGadget}
\title{Load Data via Shiny Gadget}
\usage{
loadDataGadget()
}
\description{
Use a Shiny Gadget to load the test data. Upon clicking "Load Data" a list
is returned containing the data required for the MCTestAnaylsis package.
}
