% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcTestAnalysisData.R
\name{mcTestAnalysisData}
\alias{loadAllData}
\alias{loadAnswerKey}
\alias{loadTestData}
\alias{mcTestAnalysisData}
\title{MC Test Analysis Data}
\usage{
loadTestData(mctd = NULL, test_file, has_student_id = TRUE, ...)

loadAnswerKey(mctd = NULL, answer_file, ...)

loadAllData(answer_file = NULL, test_file = NULL, has_student_id = TRUE,
  force_load = FALSE, ...)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}

\item{test_file}{Path to the test results data file}

\item{has_student_id}{Does the first column of the test data include a student identifier?}

\item{...}{Arguments passed to \code{\link{read.csv}}}

\item{answer_file}{Path to the answer key data file}

\item{force_load}{Force calculation of test analysis items (may be slow).
Default is \code{FALSE} as individual analysis will calculate needed
analysis portions on the fly when called.}
}
\description{
Load test and answer key data with \code{\link{loadTestData}} and
\code{\link{loadAnswerKey}}. Creates an MC Test Analysis data object.
}
\section{Functions}{
\itemize{
\item \code{loadTestData}: Reads test data from CSV or TSV file and applies basic preprocessing.

\item \code{loadAnswerKey}: Reads answer_key data from CSV or TSV file and applies basic preprocessing.

\item \code{loadAllData}: Read answer_key and test data from CSV or TSV
files
}}
\section{Answer Key Data Format}{

The Answer Key file should contain four columns in the following order:
\enumerate{
   \item Question number, name or identifier: Eg. \code{"Q1"}, \code{"1"}, etc.
   \item The correct answer for the question: Eg. \code{"1"}, \code{"A"}, etc.
   \item A descriptive title for the question: Eg. \code{"Taylor Function"},
         \code{"Tensor Flow"}, etc.
   \item An identifier for the concept group to which the question belongs:
         Eg. \code{"Taylor Series"}, \code{"A"}, \code{"Concept 1"}, etc.
}
}

\section{Test Data Format}{

The test results data should contain as rows each student's response, with
each question assigned a column. If the test results data contains student
identifiers, these identifiers should be included in the first column, prior
to the test answers. The MC Test Analysis Tool assumes that the question
columns are in the same order as reported in the answers data.
}

