% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTestTheory.R
\name{pbcc}
\alias{pbcc}
\title{Point-Biserial Correlation Coefficient}
\usage{
pbcc(mctd)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}
}
\description{
Calculates standard PBCC (or Pearson-moment correlation), in which the
answers to each item are correlated with overall test performance.
}

