% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTestTheory.R
\name{pbcc_modified}
\alias{pbcc_modified}
\title{Modified Point-Biserial Correlation Coefficient}
\usage{
pbcc_modified(mctd)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}
}
\description{
Calculates the modified PBCC for each item, where the item answers are
correlated with overall test performance without considering the given item
in the overall test score.
}

