% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTestTheory.R
\name{plotOverallHistogram}
\alias{plotOverallHistogram}
\title{Plot a histogram of overall test scores}
\usage{
plotOverallHistogram(mctd, add_normal = TRUE, binwidth = 0.05)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}

\item{add_normal}{If TRUE, adds a scaled normal curve to the plot.}

\item{binwidth}{Sets the \code{binwidth} of the histogram bars.}
}
\description{
Plots a histogram of overall test scores with an optional scaled normal curve
overlaid for reference.
}
