% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorAnalysis.R
\name{plotScree}
\alias{plotScree}
\title{Scree Plot}
\usage{
plotScree(mctd, return_nfactcomp = FALSE)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}

\item{return_nfactcomp}{Should suggested number of factors and components be
returned in addition to scree plot?}
}
\description{
Wraps \code{\link[psych]{fa.parallel}} to provide scree plot or automatic
determination of appropriate number of components or factors to extract.
}
