% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicTestTheory.R
\name{plotTestScoreByQuestion}
\alias{plotTestScoreByQuestion}
\title{Plot Overall Test Score by Question and Correctness}
\usage{
plotTestScoreByQuestion(mctd, concepts = unique(mctd$AnswerKey$Concept),
  facet_by_concept = FALSE)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}

\item{concepts}{Character vector containing concept groups to be plotted}

\item{facet_by_concept}{Should plot be facetted by concept group?}
}
\description{
Boxplot of test scores compared with question responses.
}
