% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemResponseTheory.R
\name{plotTetrachoric}
\alias{plotTetrachoric}
\title{Plot Tetrachoric Correlations}
\usage{
plotTetrachoric(mctd, group_by_concept = TRUE, equal_coords = FALSE)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}

\item{group_by_concept}{If \code{TRUE}, questions are grouped by concept
group and guidelines and guide text are added to the plot.}

\item{equal_coords}{If \code{TRUE}, coordinates of the plot are constrained
to be equal (square plot).}
}
\description{
This function is a simple wrapper for the \link[psych]{tetrachoric} function
provided by the \code{psych} package.
}

