% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkItemsHeuristic.R
\name{recommendItemActions}
\alias{recommendItemActions}
\title{Recommend review actions for each item}
\usage{
recommendItemActions(mctd, include_columns = c("Title", "Concept",
  "Alpha WOI", "Difficulty", "Discrimination", "PBCC"),
  digits.round = getOption("digits"))
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}
}
\description{
Recommends review actions for each item based on a set of heuristics or rules
based on published guidelines.
}

