% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{report}
\alias{report}
\alias{createReportGadget}
\alias{report}
\title{Create a report using a dialogue window}
\usage{
createReportGadget()

report()
}
\description{
Uses a Shiny Gadget to load the test data, and generate the MC Test Analysis
report. The report is saved to the current working directory and opened
automatically in the system viewer, where the user can save it to a new
directory if desired.
}
