% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorAnalysis.R
\name{summarizeEFA}
\alias{summarizeEFA}
\title{Summarize Exploratory Factor Analysis}
\usage{
summarizeEFA(mctd, cut = 0.3, ...)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}

\item{cut}{Suppress factor loadings not greater than this value}

\item{...}{Passed to \code{\link{addEFA}}.}
}
\description{
Prints a table with EFA factor loadings.
}
