% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemResponseTheory.R
\name{summarizeIRT}
\alias{summarizeIRT}
\title{Summarize IRT Model Results}
\usage{
summarizeIRT(mctd, model_params = 1,
  probcolname = "$\\\\mathrm{P}(x_i = 1 \\\\vert z = 0)$")
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}

\item{model_params}{One of \code{1, 2} or \code{3}, indicating number of 1-,
2- or 3-PL model. Can be integer or character.}

\item{probcolname}{Column name for probability column, defaults to
mathematical expression.}
}
\description{
Creates summary table of coefficients for the IRT models.
}
\seealso{
\code{\link{addIRTfits}}
}
