% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionsReview.R
\name{summarizeSelectedOptions}
\alias{summarizeSelectedOptions}
\title{Summarize Option Selection per Question}
\usage{
summarizeSelectedOptions(mctd, include_columns = c("Title", "Answer",
  "Concept"), questions_as_row_names = FALSE, as_percentage = FALSE,
  correct_vs_incorrect = FALSE)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}

\item{include_columns}{Vector of column names to include from Answer Key data
(\code{"Answer"}, \code{"Title"}, \code{"Concept"})}

\item{questions_as_row_names}{Move questions column to row name? (Default:
FALSE)}

\item{as_percentage}{Summarize by absolute count or percentage of responses}

\item{correct_vs_incorrect}{Group option into correct option vs incorrect
options? (Default: FALSE)}
}
\description{
Returns a dataframe summarizing the percentage of students who chose each of
the options availble for each question.
}
