% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemAnalysis.R
\name{testScoreByQuestionPlot}
\alias{testScoreByQuestionPlot}
\title{Plot Overall Test Score by Question and Correctness}
\usage{
testScoreByQuestionPlot(mctd, concepts = unique(mctd$AnswerKey$Concept),
  facet_by_concept = FALSE)
}
\description{
Boxplot of test scores compared with question responses.
}

