% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemResponseTheory.R
\name{addIRTfits}
\alias{addIRTfits}
\title{Add Item Response Theory Model Fits}
\usage{
addIRTfits(mctd)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}
}
\description{
Adds item response theory model fits to the \link{mcTestAnalysisData} object.
}

