% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReport.R
\name{createReport}
\alias{createReport}
\title{Create MC Test Analysis Report}
\usage{
createReport(answer_file = file.choose(), test_file = file.choose(),
  test_title = ifelse(interactive(), readline("Test title: "),
  "MC Test Analysis"), author = ifelse(interactive(),
  readline("Test author: "), Sys.info()["user"]), out_file = paste0("Report",
  " - ", test_title, ".pdf"), out_path = getwd(), open_file = TRUE,
  report_options = list(), ...)
}
\arguments{
\item{answer_file}{Path to the answer key data file}

\item{test_file}{Path to the test results data file}

\item{...}{Arguments passed to \code{\link{read.csv}}}
}
\description{
Creates a static PDF file containing a report on the properties of the test
using Classic Test Theory and Item Response Theory.
}

