% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionsReview.R
\name{optionsSelectedPct}
\alias{optionsSelectedPct}
\title{Option selection per question}
\usage{
optionsSelectedPct(mctd, include_title = TRUE,
  questions_as_row_names = FALSE, correct_vs_incorrect = FALSE)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}

\item{include_title}{Include question title column from Answer Key? (Default:
TRUE)}

\item{questions_as_row_names}{Move questions column to row name? (Default:
FALSE)}

\item{correct_vs_incorrect}{Group option into correct option vs incorrect
options? (Default: FALSE)}
}
\description{
Returns a dataframe summarizing the percentage of students who chose each of
the options availble for each question.
}

