% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemAnalysis.R
\name{summarizeCTT}
\alias{summarizeCTT}
\title{Generate Classic Test Theory Results Summary Table}
\usage{
summarizeCTT(mctd, overall_summary = TRUE)
}
\arguments{
\item{mctd}{Existing mcTestAnalysis data object}
}
\description{
Summarizes Classic Test Theory results for the loaded test.
}

